+++ grunt/concat.js
@@ -1,13 +1,13 @@
 module.exports = {
-	options: {
-      	separator: "\n/* ---- STAMPLAY JS SDK ---- */\n",
-      	stripBanners: {
-      		block: true
-      	},
-      	banner: "/*! Stamplay v<%= pkg.version %> | " + "(c) " + new Date().getFullYear() + " The Stamplay Dreamteam */"
-	},
-	dist :{
-		src: [
+      options: {
+            separator: "\n/* ---- STAMPLAY JS SDK ---- */\n",
+            stripBanners: {
+                  block: true
+            },
+            banner: "/*! Stamplay v<%= pkg.version %> | " + "(c) " + new Date().getFullYear() + " The Stamplay Dreamteam */"
+      },
+      dist: {
+            src: [
                   'bower_components/underscore/underscore-min.js',
                   'bower_components/q/q.js',
                   'bower_components/store.js/store.js',
@@ -19,8 +19,9 @@ module.exports = {
                   'src/user.js',
                   'src/cobject.js',
                   'src/webhook.js',
-                  'src/stripe.js'
+                  'src/stripe.js',
+                  'src/codeblock.js'
                 ],
-		dest: 'dist/stamplay.js'
-	}
+            dest: 'dist/stamplay.js'
+      }
 };